ETL_Pipeline/SyedAhsanAli/DS-25:2024/requirements.txt
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
gspread==6.2.0
httplib2==0.22.0
idna==3.10
numpy==2.2.4
oauthlib==3.2.2
pandas==2.2.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pymongo==4.11.3
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
tzdata==2025.2
urllib3==2.3.0
